Add intermediate base class for remote image fullscreen images (taxon + observation photos)
